Extract ViewComponent::Form::Helpers::Rails
